AIStudyCoach/requirements.txt
streamlit>=1.35.0,<2
plotly>=5.18.0,<6
pandas>=2.3.2
pytesseract>=0.3.13
googletrans==4.0.0rc1
httpcore==0.15.0
httpx==0.24.0
folium>=0.20.0
requests>=2.32.5
streamlit-folium>=0.25.1
pillow>=10.3.0
opencv-python-headless>=4.10.0
qrcode>=7.4.2
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
faster-whisper>=1.0.0
jiwer>=3.0.5
google-generativeai>=0.7.0
